core/linux-aarch64-rc to 5.12.rc3-4 (.its & mkimage regression)
With the latest mkimage from U-Boot 2021.04, the .its we have no
longer builds, failing with:

/usr/bin/mkimage: verify_header failed for FIT Image support with exit code 1

Changes and intro section above inspired by
https://lore.kernel.org/patchwork/patch/1410873/

Since we seem to be able to name these nodes whatever we want, and since
depthcharge happily prints them, let's name them more semantically:

	kernel@1 -> kernel-archlinuxarm
	fdt@12 -> sc7180-trogdor-lazor-r3-lte.dtb.lz4
	config@12 -> sc7180-trogdor-lazor-r3-lte

Delete default configuration since it's not really needed and probably
causes harm anyway if a specific dtb were to boot on a random board
(gpio mismatches and such).

The other uses .its files in the other alarm packages are also affected
due to the mkimage update. So we should update those soon as well ;)